generic get_plot_data in base.py and update prepostfit code to get hdi
